added editContactModal

+++ src/components/DeleteContactModal/DeleteContactModal.jsx
@@ -1,29 +1,48 @@
 import Modal from "react-modal";
 import { useDispatch, useSelector } from "react-redux";
+import toast from "react-hot-toast";
 import {
   selectContactToDelete,
-  selectIsModalOpen,
-} from "../../redux/modal/selectors";
-import { closeModal } from "../../redux/modal/slice";
+  selectIsDeleteModalOpen,
+} from "../../redux/contacts/selectors";
+import { closeDeleteModal } from "../../redux/contacts/slice";
 import { deleteContact } from "../../redux/contacts/operations";
+
+Modal.setAppElement("#root");
 
 export default function DeleteContactModal() {
   const dispatch = useDispatch();
-  const isModalOpen = useSelector(selectIsModalOpen);
+  const isDeleteModalOpen = useSelector(selectIsDeleteModalOpen);
   const contactToDelete = useSelector(selectContactToDelete);
 
+  // if (!isModalOpen) {
+  //   return null;
+  // }
+
   const handleCloseModal = () => {
-    dispatch(closeModal());
+    dispatch(closeDeleteModal());
   };
 
   const handleContactDelete = () => {
-    dispatch(deleteContact(contactToDelete.id));
-    dispatch(closeModal());
+    if (contactToDelete) {
+      dispatch(deleteContact(contactToDelete.id))
+        .unwrap()
+        .then(() => {
+          toast.success("Congratulations, contact deleted successfully!!!");
+        })
+
+        .catch((error) => {
+          console.log("🚀 ~ handleSubmit ~ error:", error);
+          toast.error("An error occurred, try again!!!");
+        });
+    }
+
+    dispatch(closeDeleteModal());
   };
 
   return (
     <Modal
-      isOpen={isModalOpen}
+      isOpen={isDeleteModalOpen}
       onRequestClose={handleCloseModal}
       contentLabel="Confirm Deletion"
     >
